+++ codex-ui/src/vue/components/avatar/Avatar.vue
