+++ codex-ui/src/vue/components/avatar/index.ts
