+++ codex-ui/src/vue/components/editor/index.ts
