+++ codex-ui/src/vue/index.ts
@@ -1,10 +1,10 @@
-export * from './button';
-export * from './form';
-export * from './heading';
-export * from './input';
-export * from './row';
-export * from './avatar';
-export * from './editor';
-export * from './icon';
-export * from './radio';
-export * from './switch';
+export * from './components/button';
+export * from './components/form';
+export * from './components/heading';
+export * from './components/input';
+export * from './components/row';
+export * from './components/avatar';
+export * from './components/editor';
+export * from './components/icon';
+export * from './components/radio';
+export * from './components/switch';
